requirements.txt
alabaster==0.7.12
appnope==0.1.3
asgiref==3.5.2
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.13
djangorestframework==3.13.1
docutils==0.18.1
drf-yasg==1.20.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
inflection==0.5.1
ipdb==0.13.9
ipython==7.34.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysqlclient==1.4.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
snowballstemmer==2.2.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.2.1.post0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.8.0
